+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+### Fixed
+
+- Xtheme: fix model choice widget for plugins (django 2)
+
+
 ## [2.1.2] - 2020-08-26
 
 ### Fixed