+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+
+## [2.1.5] - 2020-09-08
+
 ### Fixed
 
+- Requirements: require Markdown>=3,<4 instead <3
 - Xtheme: Fix social media plugin form initial data population.